Update pipeline with new apk-signing docker image
The server apk-signing-2.uplink.eyeo.it is being decommissioned
by the PE team.
Moving the credentials and scripts to GCP secrets manager and docker image
Adding a default runner to the Gitlab CI and overwrite the runner tag on specific jobs.

Update README with secret variable utilization section on how they are utilized,
with additional instructions on how to set up the variables for further usage.